tarball revision beecrypt-4.2.0

+++ c++/crypto/CipherSpi.cxx
@@ -28,7 +28,7 @@ using beecrypt::lang::UnsupportedOperationException;
 
 using namespace beecrypt::crypto;
 
-size_t CipherSpi::engineGetKeySize(const Key& key) const throw (InvalidKeyException)
+int CipherSpi::engineGetKeySize(const Key& key) const throw (InvalidKeyException)
 {
 	throw UnsupportedOperationException();
 }
