+++ c++/provider/SHA512Digest.cxx
@@ -55,7 +55,7 @@ const bytearray& SHA512Digest::engineDigest()
 	return _digest;
 }
 
-size_t SHA512Digest::engineDigest(byte* data, size_t offset, size_t length) throw (ShortBufferException)
+int SHA512Digest::engineDigest(byte* data, int offset, int length) throw (ShortBufferException)
 {
 	if (!data)
 		throw NullPointerException();
@@ -69,7 +69,7 @@ size_t SHA512Digest::engineDigest(byte* data, size_t offset, size_t length) thro
 	return 64;
 }
 
-size_t SHA512Digest::engineGetDigestLength()
+int SHA512Digest::engineGetDigestLength()
 {
 	return 64;
 }
@@ -86,7 +86,7 @@ void SHA512Digest::engineUpdate(byte b)
 		throw ProviderException("BeeCrypt internal error in sha512Digest");
 }
 
-void SHA512Digest::engineUpdate(const byte* data, size_t offset, size_t length)
+void SHA512Digest::engineUpdate(const byte* data, int offset, int length)
 {
 	if (sha512Update(&_param, data+offset, length))
 		throw ProviderException("BeeCrypt internal error in sha512Digest");
